# Office-management-railway-deployment/swagger.yaml
openapi: 3.0.0
info:
  title: Auto Computation Office Management API
  version: 3.0.0
  description: API documentation for the Office Management System Backend
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://office-management-sbuh.onrender.com
    description: Production server (Render)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in a cookie
security:
  - cookieAuth: []
paths:
...
  /admin/tasks/all:
    get:
      summary: Get all tasks with assignee details
      tags:
        - Admin-Tasks
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of all tasks
        "500":
          description: Server error
  /admin/tasks/create:
    post:
      summary: Create and assign a new task
      tags:
        - Admin-Tasks
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - priority
                - assigned_to
              properties:
                title:
                  type: string
                project:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum:
                    - Low
                    - Medium
                    - High
                    - Urgent
                due_date:
                  type: string
                  format: date
                assigned_to:
                  type: integer
      responses:
        "201":
          description: Task created and notification sent
        "400":
          description: Missing required fields
        "500":
          description: Server error
  /admin/tasks/{id}:
    delete:
      summary: Delete a task
      tags:
        - Admin-Tasks
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted successfully
        "500":
          description: Server error
    patch:
      summary: Update an existing task
      tags:
        - Admin-Tasks
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                project:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                status:
                  type: string
                  enum:
                    - Pending
                    - In Progress
                    - Completed
                due_date:
                  type: string
                  format: date
                assigned_to:
                  type: integer
      responses:
        "200":
          description: Task updated successfully
        "400":
          description: No fields to update
        "500":
          description: Server error
...
  /admin/meetings/create:
    post:
      summary: Schedule a new meeting
      tags:
        - Admin-Meetings
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start_time
                - end_time
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                join_url:
                  type: string
                user_ids:
                  type: array
                  items:
                    type: integer
      responses:
        "201":
          description: Meeting scheduled successfully
        "400":
          description: Missing required fields
        "500":
          description: Server error
  /admin/meetings/all:
    get:
      summary: Get all meetings
      tags:
        - Admin-Meetings
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of all meetings
        "500":
          description: Server error
  /admin/meetings/{id}:
    delete:
      summary: Cancel/Delete a meeting
      tags:
        - Admin-Meetings
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting cancelled successfully
        "500":
          description: Server error
    put:
      summary: Update an existing meeting
      tags:
        - Admin-Meetings
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start_time
                - end_time
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                join_url:
                  type: string
                user_ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Meeting updated successfully
        "404":
          description: Meeting not found
        "500":
          description: Server error
...
  /employee/tasks/all:
    get:
      summary: Get all tasks assigned to the current employee
      tags:
        - Employee-Tasks
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of assigned tasks
        "401":
          description: Unauthorized
        "500":
          description: Server error
...